+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## 2.0.0
 * 1.20.1!
+* Added `/semihardcore <player> optout <true/false>` to opt out of the lives and max health changes 
 * When out of lives, dying now instantly respawns you at the same location where you died
 * Players are now informed when max health is lost
 * Added a config option to cap the amount of max health modifier